Update README.md
This is a simple MVC ASP.NET project where traveller can make reservation with a host. A transportation provider can assist with transport if needed. Agent can step in when a traveller wants to leave or host is unable to welcome traveller. The traveller can also find out more information about each country on every continent and decide if its suitable for visit